Extract algolia indexing out of gatsby-config, upgrade algolia plugin
