+++ package.json
@@ -37,7 +37,7 @@
     "gatsby": "^2.31.1",
     "gatsby-cli": "^2.12.93",
     "gatsby-image": "^2.4.1",
-    "gatsby-plugin-algolia": "^0.13.0",
+    "gatsby-plugin-algolia": "^0.16.3",
     "gatsby-plugin-catch-links": "^2.6.0",
     "gatsby-plugin-google-fonts": "^1.0.1",
     "gatsby-plugin-google-tagmanager": "^2.10.0",
